feat: Add schedulecontrollers.js
- Added schedulecontrollers.js to handle the logic for the schedule routes
- Updated scheduleroutes.js to use the new controller
