+++ routes/scheduleroutes.js
@@ -1,88 +1,18 @@
 const express = require('express');
-const conn = require('../utils/dbconn');
+const controller = require('./../controllers/schedulecontrollers');
 const router = express.Router();
 
-router.get('/', (req, res) => {
 
-    const selectSQL = `SELECT * FROM runschedule`;
-    conn.query(selectSQL, (err, rows) => {
-        if (err) {
-            throw err;
-        } else {
-            res.render('index', { schedule: rows });
-        }
-    });
-});
+router.get('/', controller.getSchedule);
 
-router.get('/new', (req, res) => {
-    res.render('addschedule');
-});
+router.get('/new', controller.getAddNewRun);
 
+router.get('/edit/:id', controller.selectRun);
 
-router.post('/new', (req, res) => {
-    const { new_details, new_date } = req.body;
-    const vals = [new_details, new_date];
+router.post('/new', controller.postNewRun);
 
-    const insertSQL = `INSERT INTO runschedule (items, mydate) VALUES (?, ?)`;
+router.post('/edit/:id', controller.updateRun);
 
-    conn.query(insertSQL, vals, (err, rows) => {
-        if (err) {
-            throw err;
-        } else {
-            res.redirect('/');
-        }
-    });
-
-});
-
-router.get('/edit/:id', (req, res) => {
-    const { id } = req.params;
-
-    const selectSQL = `SELECT * FROM runschedule WHERE id = ${id}`;
-    conn.query(selectSQL, (err, rows) => {
-        if (err) {
-            throw err;
-        } else {
-            console.log(rows);
-            res.render('editschedule', { details: rows });
-        }
-    });
-});
-
-router.post('/edit/:id', (req, res) => {
-    console.log("req.params",req.params);
-    console.log("req.body",req.body);
-
-
-    const run_id = req.params.id;
-    const { run_details, run_date } = req.body;
-
-    // Log the received data
-    console.log('Received data:', { run_details, run_date });
-
-
-    const vals = [run_details, run_date, run_id];
-    console.log(vals);
-
-    const updateSQL = 'UPDATE runschedule SET items = ?, mydate = ? WHERE id = ?';
-
-    conn.query(updateSQL, vals, (err, rows) => {
-        if (err) {
-            throw err;
-        } else {
-            res.redirect('/');
-        }
-    });
-});
-
-router.post('/del/:id', (req, res) => {
-    const run_id = req.params.id;
-const deleteSQL = `DELETE FROM runschedule WHERE id = ${run_id}`; conn.query(deleteSQL, (err, rows) => {
-        if (err) {
-            throw err;
-} else {
-            res.redirect('/');
-        }
-}); });
+router.post('/del/:id', controller.deleteRun);
 
 module.exports = router;
